requirements.txt
autopep8==1.5.7
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
googletrans==3.1.0a0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
psycopg2==2.9.1
pycodestyle==2.8.0
pyTelegramBotAPI==4.1.1
python-dotenv==0.19.1
requests==2.26.0
rfc3986==1.5.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
